SJF-002 Navigate betwen screens
