+++ package.json
@@ -17,8 +17,10 @@
     "react": "16.9.0",
     "react-native": "0.61.4",
     "react-native-elements": "^1.2.7",
+    "react-native-gesture-handler": "^1.5.2",
     "react-native-modal-datetime-picker": "^8.0.3",
-    "react-navigation": "^4.0.10"
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3"
   },
   "devDependencies": {
     "@babel/core": "^7.7.2",
